# crm-dashboard/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  description: |
    API for managing church members, cell groups (CGF), attendance, ministries, and events.
    This API provides CRUD operations for all entities and dashboard analytics endpoints.
  version: 1.0.0
  contact:
    name: Church Admin
    email: admin@church.org
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.church.org/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /status-history:
    get:
      tags: [Status]
      summary: Get all status history
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: no_jemaat
          in: query
          description: Filter by member ID
          schema:
            type: integer
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [Active, Inactive, Sabbatical, Moved]
        - name: start_date
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistoryListResponse'

    post:
      tags: [Status]
      summary: Create status change record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusHistoryCreate'
      responses:
        '201':
          description: Status history created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /status-history/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [Status]
      summary: Get status history by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Status]
      summary: Update status history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusHistoryUpdate'
      responses:
        '200':
          description: Status history updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Status]
      summary: Delete status history
      responses:
        '200':
          description: Status history deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'
  /members/{no_jemaat}/status-history:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [Status, Members]
      summary: Get status history for a member
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistoryListResponse'

  # ==================== EVENTS ENDPOINTS ====================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== PARAMETERS ====================
  parameters:
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    MemberId:
      name: no_jemaat
      in: path
      required: true
      description: Member number (primary key)
      schema:
        type: integer
        minimum: 1

  # ==================== RESPONSES ====================
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 400
              message: "Validation error"
              details:
                - field: "nama_jemaat"
                  message: "Name is required"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 404
              message: "Resource not found"
...
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
              example: 400
            message:
              type: string
              example: "Error message"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10

    # Member Schemas
    Member:
      type: object
      properties:
        no_jemaat:
          type: integer
          description: Member number (primary key)
          example: 1
        nama_jemaat:
          type: string
          example: "John Doe"
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
          example: "Laki-laki"
        tanggal_lahir:
          type: string
          format: date
          example: "1990-05-15"
        tahun_lahir:
          type: integer
          example: 1990
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
          example: 5
        kuliah_kerja:
          type: string
          example: "Software Engineer"
        no_handphone:
          type: string
          example: "081234567890"
        ketertarikan_cgf:
          type: string
          example: "Yes"
        nama_cgf:
          type: string
          example: "Grace CGF"
        kategori_domisili:
          type: string
          example: "Jakarta"
        alamat_domisili:
          type: string
          example: "Jl. Sudirman No. 123"
...
    StatusHistory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        no_jemaat:
          type: integer
          example: 1
        status:
          type: string
          enum: [Active, Inactive, Sabbatical, Moved]
          example: "Active"
        changed_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        reason:
          type: string
          example: "Initial registration"
        member:
          $ref: '#/components/schemas/Member'
    StatusHistoryCreate:
      type: object
      required: [no_jemaat, status]
      properties:
        no_jemaat:
          type: integer
        status:
          type: string
          enum: [Active, Inactive, Sabbatical, Moved]
        reason:
          type: string
    StatusHistoryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [Active, Inactive, Sabbatical, Moved]
        reason:
          type: string
    StatusHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/StatusHistory'
    StatusHistoryListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistory'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Event Schemas
...